requirements.txt
asgiref==3.5.0
attrs==21.4.0
autobahn==22.3.2
Automat==20.2.0
cffi==1.15.0
channels==3.0.4
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
Django==4.0.4
django-nested-admin==3.4.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
python-monkey-business==1.0.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing-extensions==4.2.0
zope.interface==5.4.0
